+++ crates/devela/src/sys/hw/pin/i2c/_.rs
@@ -4,6 +4,7 @@
 
 crate::mods_in! {
     mod addr;
+    #[cfg(all(feature = "unsafe_mmio", not(feature = "safe_sys")))]
     mod control;
     mod error;
     mod write;
@@ -12,9 +13,10 @@ crate::mods_out! { // _mods
     _mods {
         pub use super::{
             addr::I2cAddr7,
-            control::{I2cControl, I2cController},
             error::I2cError,
             write::I2cWrite,
         };
+        #[cfg(all(feature = "unsafe_mmio", not(feature = "safe_sys")))]
+        pub use super::control::{I2cControl, I2cController};
     }
 }