+++ crates/micros/src/device/_.rs
@@ -6,26 +6,40 @@
 //
 
 crate::mods_in! {
-    // mod_ audio;   // Audio converters, amplifiers and interface devices
-    // mod_ comm;    // Communication transceivers, modems, and interface devices
-    // mod_ display; // Display controllers and display-device drivers
-    // mod_ input;   // Input controllers and human-interface devices
-    // mod_ motion;  // Motor, actuator, and motion-control devices
-    // mod_ power;   // Power-management, conversion, and control devices
-    // mod_ sensor;  // Sensors and measurement devices
-    // mod_ vision;  // Cameras, imagers, and vision-oriented devices
+    // #[cfg_attr(not(nightly_doc), cfg(feature = "audio"))]
+    // pub mod_ audio;   // Audio converters, amplifiers and interface devices
+    // #[cfg_attr(not(nightly_doc), cfg(feature = "comm"))]
+    // pub mod_ comm;    // Communication transceivers, modems, and interface devices
+    #[cfg_attr(not(nightly_doc), cfg(feature = "display"))]
+    pub mod_ display; // Display controllers and display-device drivers
+    // #[cfg_attr(not(nightly_doc), cfg(feature = "input"))]
+    // pub mod_ input;   // Input controllers and human-interface devices
+    // #[cfg_attr(not(nightly_doc), cfg(feature = "motion"))]
+    // pub mod_ motion;  // Motor, actuator, and motion-control devices
+    // #[cfg_attr(not(nightly_doc), cfg(feature = "power"))]
+    // pub mod_ power;   // Power-management, conversion, and control devices
+    // #[cfg_attr(not(nightly_doc), cfg(feature = "sensor"))]
+    // pub mod_ sensor;  // Sensors and measurement devices
+    // #[cfg_attr(not(nightly_doc), cfg(feature = "vision"))]
+    // pub mod_ vision;  // Cameras, imagers, and vision-oriented devices
 }
-crate::mods_out! { // _mods
-    _mods {
-        // pub use super::{
-        //     // audio::_all::*,
-        //     // comm::_all::*,
-        //     // display::_all::*,
-        //     // input::_all::*,
-        //     // motion::_all::*,
-        //     // power::_all::*,
-        //     // sensor::_all::*,
-        //     // vision::_all::*,
-        // };
+crate::mods_out! { // _pub_mods
+    _pub_mods {
+        // #[cfg_attr(not(nightly_doc), cfg(feature = "audio"))]
+        // pub use super::audio::_all::*;
+        // #[cfg_attr(not(nightly_doc), cfg(feature = "comm"))]
+        // pub use super::comm::_all::*;
+        #[cfg_attr(not(nightly_doc), cfg(feature = "display"))]
+        pub use super::display::_all::*;
+        // #[cfg_attr(not(nightly_doc), cfg(feature = "input"))]
+        // pub use super::input::_all::*;
+        // #[cfg_attr(not(nightly_doc), cfg(feature = "motion"))]
+        // pub use super::motion::_all::*;
+        // #[cfg_attr(not(nightly_doc), cfg(feature = "power"))]
+        // pub use super::power::_all::*;
+        // #[cfg_attr(not(nightly_doc), cfg(feature = "sensor"))]
+        // pub use super::sensor::_all::*;
+        // #[cfg_attr(not(nightly_doc), cfg(feature = "vision"))]
+        // pub use super::vision::_all::*;
     }
 }